IIWA-35: Final (please!) creation of models to fit new db designs.

+++ src/backend-api/Models/Computer.cs
@@ -12,14 +12,14 @@ namespace backend_api.Models
         public int? Ssdgb { get; set; }
         public DateTime? PurchaseDate { get; set; }
         public DateTime? RenewalDate { get; set; }
-        public double? FlatCost { get; set; }
+        public decimal? FlatCost { get; set; }
         public string MonitorOutput { get; set; }
         public DateTime? EndOfLife { get; set; }
         public int? EmployeeId { get; set; }
         public bool IsAssigned { get; set; }
         public string TextField { get; set; }
-        public double? ScreenSize { get; set; }
-        public double? CostPerYear { get; set; }
+        public decimal? ScreenSize { get; set; }
+        public decimal? CostPerYear { get; set; }
         public bool IsDeleted { get; set; }
         public int? Resolution { get; set; }
