+++ src/backend-api/Models/Peripheral.cs
@@ -11,9 +11,9 @@ namespace backend_api.Models
         public string TextField { get; set; }
         public int? EmployeeId { get; set; }
         public bool IsAssigned { get; set; }
-        public double? FlatCost { get; set; }
+        public decimal? FlatCost { get; set; }
         public DateTime? PurchaseDate { get; set; }
-        public double? CostPerYear { get; set; }
+        public decimal? CostPerYear { get; set; }
         public bool IsDeleted { get; set; }
 
         public Employee Employee { get; set; }
