+++ src/backend-api/Models/Server.cs
@@ -15,11 +15,11 @@ namespace backend_api.Models
         public DateTime? RenewalDate { get; set; }
         public int? EmployeeId { get; set; }
         public DateTime? PurchaseDate { get; set; }
-        public double? FlatCost { get; set; }
+        public decimal? FlatCost { get; set; }
         public DateTime? EndOfLife { get; set; }
         public bool IsAssigned { get; set; }
         public string TextField { get; set; }
-        public double? CostPerYear { get; set; }
+        public decimal? CostPerYear { get; set; }
         public bool IsDeleted { get; set; }
 
         public Employee Employee { get; set; }
